Moved util types to internal
